update About SSW popup
