+++ src/MobileUI/PopupPages/AboutSswPage.xaml.cs
@@ -1,0 +1,39 @@
+﻿using CommunityToolkit.Maui.Behaviors;
+using CommunityToolkit.Maui.Core;
+using Mopups.Services;
+
+namespace SSW.Rewards.PopupPages;
+
+public partial class AboutSswPage
+{
+    private readonly Color _parentPageStatusBarColor;
+
+    public AboutSswPage(Color parentPageStatusBarColor = null)
+    {
+        _parentPageStatusBarColor = parentPageStatusBarColor ?? Colors.Black;
+        InitializeComponent();
+    }
+
+    private async void Handle_CloseTapped(object sender, EventArgs args)
+    {
+        await MopupService.Instance.PopAllAsync();
+    }
+
+    private async void FindoutMore_Tapped(object sender, EventArgs e)
+    {
+        await Browser.OpenAsync("https://www.ssw.com.au/ssw/Company/AboutUs.aspx", BrowserLaunchMode.External);
+    }
+
+    protected override void OnDisappearing()
+    {
+        base.OnDisappearing();
+
+        // Change status bar back
+        this.Behaviors.Clear();
+        this.Behaviors.Add(new StatusBarBehavior
+        {
+            StatusBarColor = _parentPageStatusBarColor,
+            StatusBarStyle = StatusBarStyle.LightContent,
+        });
+    }
+}
