+++ src/MobileUI/PopupPages/CameraPage.xaml.cs
@@ -5,18 +5,10 @@ namespace SSW.Rewards.PopupPages;
 
 public partial class CameraPage
 {
-    public async void Handle_Tapped(object sender, EventArgs e)
-    {
-        await MopupService.Instance.PopAllAsync();
-    }
-
-    private CameraPageViewModel _viewModel { get; set; }
-
     public CameraPage(CameraPageViewModel cameraPageViewModel)
     {
         InitializeComponent();
-        _viewModel = cameraPageViewModel;
-        _viewModel.Navigation = Navigation;
-        BindingContext = _viewModel;
+        cameraPageViewModel.Navigation = Navigation;
+        BindingContext = cameraPageViewModel;
     }
 }
